fix: add ConnectorAuditProducerIT test module
