+++ activiti-cloud-runtime-bundle-service/activiti-cloud-starter-runtime-bundle/src/test/java/org/activiti/cloud/starter/tests/services/audit/AuditProducerIT.java
@@ -5,8 +5,6 @@ import static org.activiti.api.model.shared.event.VariableEvent.VariableEvents.V
 import static org.activiti.api.process.model.events.BPMNActivityEvent.ActivityEvents.ACTIVITY_CANCELLED;
 import static org.activiti.api.process.model.events.BPMNActivityEvent.ActivityEvents.ACTIVITY_COMPLETED;
 import static org.activiti.api.process.model.events.BPMNActivityEvent.ActivityEvents.ACTIVITY_STARTED;
-import static org.activiti.api.process.model.events.BPMNErrorReceivedEvent.ErrorEvents.ERROR_RECEIVED;
-import static org.activiti.api.process.model.events.IntegrationEvent.IntegrationEvents.INTEGRATION_ERROR_RECEIVED;
 import static org.activiti.api.process.model.events.IntegrationEvent.IntegrationEvents.INTEGRATION_REQUESTED;
 import static org.activiti.api.process.model.events.IntegrationEvent.IntegrationEvents.INTEGRATION_RESULT_RECEIVED;
 import static org.activiti.api.process.model.events.ProcessDefinitionEvent.ProcessDefinitionEvents.PROCESS_DEPLOYED;
@@ -45,9 +43,6 @@ import java.util.stream.Stream;
 
 import org.activiti.api.model.shared.event.RuntimeEvent;
 import org.activiti.api.model.shared.model.ApplicationElement;
-import org.activiti.api.process.model.BPMNActivity;
-import org.activiti.api.process.model.BPMNError;
-import org.activiti.api.process.model.IntegrationContext;
 import org.activiti.api.process.model.ProcessInstance;
 import org.activiti.api.process.model.builders.ProcessPayloadBuilder;
 import org.activiti.api.process.model.builders.StartProcessPayloadBuilder;
@@ -58,21 +53,12 @@ import org.activiti.api.task.model.TaskCandidateGroup;
 import org.activiti.api.task.model.TaskCandidateUser;
 import org.activiti.api.task.model.builders.TaskPayloadBuilder;
 import org.activiti.cloud.api.model.shared.events.CloudRuntimeEvent;
-import org.activiti.cloud.api.process.model.CloudBpmnError;
 import org.activiti.cloud.api.process.model.CloudProcessDefinition;
 import org.activiti.cloud.api.process.model.CloudProcessInstance;
-import org.activiti.cloud.api.process.model.events.CloudBPMNActivityCompletedEvent;
 import org.activiti.cloud.api.process.model.events.CloudBPMNActivityStartedEvent;
-import org.activiti.cloud.api.process.model.events.CloudBPMNErrorReceivedEvent;
-import org.activiti.cloud.api.process.model.events.CloudIntegrationErrorReceivedEvent;
-import org.activiti.cloud.api.process.model.events.CloudIntegrationRequestedEvent;
-import org.activiti.cloud.api.process.model.events.CloudIntegrationResultReceivedEvent;
 import org.activiti.cloud.api.process.model.events.CloudProcessDeployedEvent;
 import org.activiti.cloud.api.process.model.impl.CandidateGroup;
 import org.activiti.cloud.api.process.model.impl.CandidateUser;
-import org.activiti.cloud.api.process.model.impl.IntegrationErrorImpl;
-import org.activiti.cloud.api.process.model.impl.IntegrationRequestImpl;
-import org.activiti.cloud.api.process.model.impl.IntegrationResultImpl;
 import org.activiti.cloud.api.task.model.CloudTask;
 import org.activiti.cloud.api.task.model.events.CloudTaskCancelledEvent;
 import org.activiti.cloud.api.task.model.events.CloudTaskCandidateUserRemovedEvent;
@@ -86,7 +72,6 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.web.client.TestRestTemplate;
-import org.springframework.cloud.stream.binding.BinderAwareChannelResolver;
 import org.springframework.core.ParameterizedTypeReference;
 import org.springframework.hateoas.PagedResources;
 import org.springframework.hateoas.Resource;
@@ -94,8 +79,6 @@ import org.springframework.hateoas.Resources;
 import org.springframework.http.HttpMethod;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.messaging.MessageChannel;
-import org.springframework.messaging.support.MessageBuilder;
 import org.springframework.test.annotation.DirtiesContext;
 import org.springframework.test.context.ActiveProfiles;
 import org.springframework.test.context.ContextConfiguration;
@@ -136,15 +119,6 @@ public class AuditProducerIT {
 
     @Autowired
     private AuditConsumerStreamHandler streamHandler;
-
-    @Autowired
-    private BinderAwareChannelResolver channelResolver;
-
-    @Value("integrationResult_${spring.application.name}")
-    private String integrationResultDestination;
-
-    @Value("integrationError_${spring.application.name}")
-    private String integrationErrorDestination;
 
     private Map<String, String> processDefinitionIds = new HashMap<>();
 
@@ -904,362 +878,6 @@ public class AuditProducerIT {
                     .contains(
                             tuple(TASK_CANCELLED, "My Task 3"),
                             tuple(TASK_CANCELLED, "My Task 4")
-                    );
-
-            assertThat(receivedEvents).extracting(RuntimeEvent::getEventType,
-                    RuntimeEvent::getProcessInstanceId)
-                    .contains(tuple(PROCESS_COMPLETED, startProcessEntity.getBody().getId()));
-        });
-    }
-
-    @Test
-    public void shouldProduceIntegrationResultEventsDuringMultiInstanceCloudConnectorExecution() {
-
-        //when
-        ResponseEntity<CloudProcessInstance> startProcessEntity = processInstanceRestTemplate.startProcessByKey("miParallelCloudConnector",
-                                                                                                                Collections.singletonMap("instanceCount", 3),
-                                                                                                                null);
-
-        List<CloudIntegrationRequestedEvent> integrationRequestedEvents = new ArrayList<>();
-
-        //then
-        await()
-            .untilAsserted(() -> {
-            List<CloudRuntimeEvent<?, ?>> receivedEvents = streamHandler.getAllReceivedEvents();
-            receivedEvents = receivedEvents.stream()
-                    .filter(event -> startProcessEntity.getBody().getId().equals(event.getProcessInstanceId()))
-                    .collect(Collectors.toList());
-
-            assertThat(streamHandler.getReceivedHeaders()).containsKeys(ALL_REQUIRED_HEADERS);
-
-
-            List<CloudBPMNActivityStartedEvent> receivedActivityStartedEvents = receivedEvents.stream()
-                    .filter(event -> event.getEventType() == ACTIVITY_STARTED && event.getEntityId()
-                                                                                      .equals("miCloudConnectorId"))
-                    .map(CloudBPMNActivityStartedEvent.class::cast)
-                    .collect(Collectors.toList());
-
-            assertThat(receivedActivityStartedEvents)
-                    .extracting(RuntimeEvent::getEventType, event -> ((BPMNActivity) event.getEntity()).getElementId())
-                    .containsExactlyInAnyOrder(
-                            tuple(ACTIVITY_STARTED, "miCloudConnectorId"),
-                            tuple(ACTIVITY_STARTED, "miCloudConnectorId"),
-                            tuple(ACTIVITY_STARTED, "miCloudConnectorId")
-                    );
-
-            List<CloudIntegrationRequestedEvent> receivedIntegrationRequestedEvents = receivedEvents.stream()
-                    .filter(event -> event.getEventType() == INTEGRATION_REQUESTED &&
-                                        ((IntegrationContext) event.getEntity()).getClientId()
-                                                                                .equals("miCloudConnectorId"))
-                    .map(CloudIntegrationRequestedEvent.class::cast)
-                    .collect(Collectors.toList());
-
-            assertThat(receivedIntegrationRequestedEvents)
-                    .extracting(RuntimeEvent::getEventType, event -> ((IntegrationContext) event.getEntity()).getClientId())
-                    .containsExactlyInAnyOrder(
-                            tuple(INTEGRATION_REQUESTED, "miCloudConnectorId"),
-                            tuple(INTEGRATION_REQUESTED, "miCloudConnectorId"),
-                            tuple(INTEGRATION_REQUESTED, "miCloudConnectorId")
-                    );
-
-            integrationRequestedEvents.addAll(receivedIntegrationRequestedEvents);
-
-        });
-
-        MessageChannel resultsChannel = channelResolver.resolveDestination(integrationResultDestination);
-
-        // complete cloud connector tasks
-        integrationRequestedEvents.stream()
-                                  .map(request -> {
-                                      return new IntegrationResultImpl(new IntegrationRequestImpl(request.getEntity()),
-                                                                       request.getEntity());
-                                  })
-                                  .map(payload -> MessageBuilder.withPayload(payload)
-                                                                .build())
-                                  .forEach(resultsChannel::send);
-
-
-        //then
-        await()
-            .untilAsserted(() -> {
-            List<CloudRuntimeEvent<?, ?>> receivedEvents = streamHandler.getAllReceivedEvents();
-            receivedEvents = receivedEvents.stream()
-                    .filter(event -> startProcessEntity.getBody().getId().equals(event.getProcessInstanceId()))
-                    .collect(Collectors.toList());
-
-            List<CloudBPMNActivityCompletedEvent> receivedActivityCompletedEvents = receivedEvents.stream()
-                    .filter(event -> event.getEventType() == ACTIVITY_COMPLETED && event.getEntityId()
-                                                                                      .equals("miCloudConnectorId"))
-                    .map(CloudBPMNActivityCompletedEvent.class::cast)
-                    .collect(Collectors.toList());
-
-            assertThat(receivedActivityCompletedEvents)
-                    .extracting(RuntimeEvent::getEventType, event -> ((BPMNActivity) event.getEntity()).getElementId())
-                    .containsExactlyInAnyOrder(
-                            tuple(ACTIVITY_COMPLETED, "miCloudConnectorId"),
-                            tuple(ACTIVITY_COMPLETED, "miCloudConnectorId"),
-                            tuple(ACTIVITY_COMPLETED, "miCloudConnectorId")
-                    );
-
-            List<CloudIntegrationResultReceivedEvent> receivedIntegrationResultEvents = receivedEvents.stream()
-                    .filter(event -> event.getEventType() == INTEGRATION_RESULT_RECEIVED &&
-                                        ((IntegrationContext) event.getEntity()).getClientId()
-                                                                                .equals("miCloudConnectorId"))
-                    .map(CloudIntegrationResultReceivedEvent.class::cast)
-                    .collect(Collectors.toList());
-
-            assertThat(receivedIntegrationResultEvents)
-                    .extracting(RuntimeEvent::getEventType, event -> ((IntegrationContext) event.getEntity()).getClientId())
-                    .containsExactlyInAnyOrder(
-                            tuple(INTEGRATION_RESULT_RECEIVED, "miCloudConnectorId"),
-                            tuple(INTEGRATION_RESULT_RECEIVED, "miCloudConnectorId"),
-                            tuple(INTEGRATION_RESULT_RECEIVED, "miCloudConnectorId")
-                    );
-
-            assertThat(receivedEvents).extracting(RuntimeEvent::getEventType,
-                    RuntimeEvent::getProcessInstanceId)
-                    .contains(tuple(PROCESS_COMPLETED, startProcessEntity.getBody().getId()));
-        });
-    }
-
-    @Test
-    public void shouldProduceIntegrationErrorEventsDuringMultiInstanceCloudConnectorExecution() {
-
-        //when
-        ResponseEntity<CloudProcessInstance> startProcessEntity = processInstanceRestTemplate.startProcessByKey("miParallelCloudConnector",
-                                                                                                                Collections.singletonMap("instanceCount", 3),
-                                                                                                                null);
-        List<CloudIntegrationRequestedEvent> integrationRequestedEvents = new ArrayList<>();
-
-        //then
-        await()
-            .untilAsserted(() -> {
-            List<CloudRuntimeEvent<?, ?>> receivedEvents = streamHandler.getAllReceivedEvents();
-            receivedEvents = receivedEvents.stream()
-                    .filter(event -> startProcessEntity.getBody().getId().equals(event.getProcessInstanceId()))
-                    .collect(Collectors.toList());
-
-            assertThat(streamHandler.getReceivedHeaders()).containsKeys(ALL_REQUIRED_HEADERS);
-
-
-            List<CloudBPMNActivityStartedEvent> receivedActivityStartedEvents = receivedEvents.stream()
-                    .filter(event -> event.getEventType() == ACTIVITY_STARTED && event.getEntityId()
-                                                                                      .equals("miCloudConnectorId"))
-                    .map(CloudBPMNActivityStartedEvent.class::cast)
-                    .collect(Collectors.toList());
-
-            assertThat(receivedActivityStartedEvents)
-                    .extracting(RuntimeEvent::getEventType, event -> ((BPMNActivity) event.getEntity()).getElementId())
-                    .containsExactlyInAnyOrder(
-                            tuple(ACTIVITY_STARTED, "miCloudConnectorId"),
-                            tuple(ACTIVITY_STARTED, "miCloudConnectorId"),
-                            tuple(ACTIVITY_STARTED, "miCloudConnectorId")
-                    );
-
-            List<CloudIntegrationRequestedEvent> receivedIntegrationRequestedEvents = receivedEvents.stream()
-                    .filter(event -> event.getEventType() == INTEGRATION_REQUESTED &&
-                                        ((IntegrationContext) event.getEntity()).getClientId()
-                                                                                .equals("miCloudConnectorId"))
-                    .map(CloudIntegrationRequestedEvent.class::cast)
-                    .collect(Collectors.toList());
-
-            assertThat(receivedIntegrationRequestedEvents)
-                    .extracting(RuntimeEvent::getEventType, event -> ((IntegrationContext) event.getEntity()).getClientId())
-                    .containsExactlyInAnyOrder(
-                            tuple(INTEGRATION_REQUESTED, "miCloudConnectorId"),
-                            tuple(INTEGRATION_REQUESTED, "miCloudConnectorId"),
-                            tuple(INTEGRATION_REQUESTED, "miCloudConnectorId")
-                    );
-
-            integrationRequestedEvents.addAll(receivedIntegrationRequestedEvents);
-
-        });
-
-        MessageChannel resultsChannel = channelResolver.resolveDestination(integrationErrorDestination);
-
-        Error error = new Error("IntegrationError");
-        error.fillInStackTrace();
-
-        // throw error in cloud connector
-        integrationRequestedEvents.stream()
-                                  .map(request -> {
-                                      return new IntegrationErrorImpl(new IntegrationRequestImpl(request.getEntity()),
-                                                                      error);
-                                  })
-                                  .map(payload -> MessageBuilder.withPayload(payload)
-                                                                .build())
-                                  .forEach(resultsChannel::send);
-        //then
-        await()
-            .untilAsserted(() -> {
-            List<CloudRuntimeEvent<?, ?>> receivedEvents = streamHandler.getAllReceivedEvents();
-            receivedEvents = receivedEvents.stream()
-                    .filter(event -> startProcessEntity.getBody().getId().equals(event.getProcessInstanceId()))
-                    .collect(Collectors.toList());
-
-            List<CloudIntegrationErrorReceivedEvent> receivedIntegrationResultEvents = receivedEvents.stream()
-                    .filter(event -> event.getEventType() == INTEGRATION_ERROR_RECEIVED &&
-                                        ((IntegrationContext) event.getEntity()).getClientId()
-                                                                                .equals("miCloudConnectorId"))
-                    .map(CloudIntegrationErrorReceivedEvent.class::cast)
-                    .collect(Collectors.toList());
-
-            assertThat(receivedIntegrationResultEvents)
-                    .extracting(RuntimeEvent::getEventType, event -> ((IntegrationContext) event.getEntity()).getClientId())
-                    .containsExactlyInAnyOrder(
-                            tuple(INTEGRATION_ERROR_RECEIVED, "miCloudConnectorId"),
-                            tuple(INTEGRATION_ERROR_RECEIVED, "miCloudConnectorId"),
-                            tuple(INTEGRATION_ERROR_RECEIVED, "miCloudConnectorId")
-                    );
-        });
-    }
-
-    @Test
-    public void shouldProduceIntegrationCloudBpmnErrorEventsForCloudBpmnErrorConnectorProcess() {
-
-        //when
-        ResponseEntity<CloudProcessInstance> startProcessEntity = processInstanceRestTemplate.startProcessByKey("cloudBpmnErrorCloudConnectorProcess",
-                                                                                                                null,
-                                                                                                                null);
-        List<CloudIntegrationRequestedEvent> integrationRequestedEvents = new ArrayList<>();
-
-        //then
-        await()
-            .untilAsserted(() -> {
-            List<CloudRuntimeEvent<?, ?>> receivedEvents = streamHandler.getAllReceivedEvents();
-            receivedEvents = receivedEvents.stream()
-                    .filter(event -> startProcessEntity.getBody().getId().equals(event.getProcessInstanceId()))
-                    .collect(Collectors.toList());
-
-            assertThat(streamHandler.getReceivedHeaders()).containsKeys(ALL_REQUIRED_HEADERS);
-
-            List<CloudIntegrationRequestedEvent> receivedIntegrationRequestedEvents = receivedEvents.stream()
-                    .filter(event -> event.getEventType() == INTEGRATION_REQUESTED)
-                    .map(CloudIntegrationRequestedEvent.class::cast)
-                    .collect(Collectors.toList());
-
-            assertThat(receivedIntegrationRequestedEvents)
-                    .extracting(RuntimeEvent::getEventType, event -> ((IntegrationContext) event.getEntity()).getClientId())
-                    .containsExactlyInAnyOrder(
-                            tuple(INTEGRATION_REQUESTED, "performBusinessTaskCloudConnector")
-                    );
-
-            integrationRequestedEvents.addAll(receivedIntegrationRequestedEvents);
-
-        });
-
-        MessageChannel errorChannel = channelResolver.resolveDestination(integrationErrorDestination);
-
-        CloudBpmnError error = new CloudBpmnError("CLOUD_BPMN_ERROR");
-
-        // when throw error in cloud connector
-        integrationRequestedEvents.stream()
-                                  .map(request -> {
-                                      return new IntegrationErrorImpl(new IntegrationRequestImpl(request.getEntity()),
-                                                                      error);
-                                  })
-                                  .map(payload -> MessageBuilder.withPayload(payload)
-                                                                .build())
-                                  .forEach(errorChannel::send);
-        //then
-        await()
-            .untilAsserted(() -> {
-            List<CloudRuntimeEvent<?, ?>> receivedEvents = streamHandler.getAllReceivedEvents();
-            receivedEvents = receivedEvents.stream()
-                    .filter(event -> startProcessEntity.getBody().getId().equals(event.getProcessInstanceId()))
-                    .collect(Collectors.toList());
-
-            List<CloudIntegrationErrorReceivedEvent> receivedIntegrationResultEvents = receivedEvents.stream()
-                    .filter(event -> event.getEventType() == INTEGRATION_ERROR_RECEIVED)
-                    .map(CloudIntegrationErrorReceivedEvent.class::cast)
-                    .collect(Collectors.toList());
-
-            assertThat(receivedIntegrationResultEvents)
-                    .extracting(RuntimeEvent::getEventType, event -> ((IntegrationContext) event.getEntity()).getClientId())
-                    .containsExactlyInAnyOrder(
-                            tuple(INTEGRATION_ERROR_RECEIVED, "performBusinessTaskCloudConnector")
-                    );
-
-            List<CloudBPMNErrorReceivedEvent> receivedBmpnErrorEvents = receivedEvents.stream()
-                    .filter(event -> event.getEventType() == ERROR_RECEIVED)
-                    .map(CloudBPMNErrorReceivedEvent.class::cast)
-                    .collect(Collectors.toList());
-
-            assertThat(receivedBmpnErrorEvents)
-                    .extracting(CloudBPMNErrorReceivedEvent::getEventType,
-                                event -> ((BPMNError) event.getEntity()).getErrorCode())
-                    .containsExactlyInAnyOrder(
-                            tuple(ERROR_RECEIVED,
-                                  "CLOUD_BPMN_ERROR")
-                    );
-        });
-
-        // given reset state
-        integrationRequestedEvents.clear();
-        streamHandler.clear();
-
-        // when fix business error
-        List<CloudTask> tasks = new ArrayList<>(processInstanceRestTemplate.getTasks(startProcessEntity).getBody().getContent());
-        assertThat(tasks).hasSize(1);
-
-        taskRestTemplate.complete(tasks.get(0));
-
-        // then
-        await()
-            .untilAsserted(() -> {
-            List<CloudRuntimeEvent<?, ?>> receivedEvents = streamHandler.getAllReceivedEvents();
-            receivedEvents = receivedEvents.stream()
-                    .filter(event -> startProcessEntity.getBody().getId().equals(event.getProcessInstanceId()))
-                    .collect(Collectors.toList());
-
-            assertThat(streamHandler.getReceivedHeaders()).containsKeys(ALL_REQUIRED_HEADERS);
-
-
-            List<CloudIntegrationRequestedEvent> receivedIntegrationRequestedEvents = receivedEvents.stream()
-                    .filter(event -> event.getEventType() == INTEGRATION_REQUESTED)
-                    .map(CloudIntegrationRequestedEvent.class::cast)
-                    .collect(Collectors.toList());
-
-            assertThat(receivedIntegrationRequestedEvents)
-                    .extracting(RuntimeEvent::getEventType, event -> ((IntegrationContext) event.getEntity()).getClientId())
-                    .containsExactlyInAnyOrder(
-                            tuple(INTEGRATION_REQUESTED, "performBusinessTaskCloudConnector")
-                    );
-
-            integrationRequestedEvents.addAll(receivedIntegrationRequestedEvents);
-
-        });
-
-        MessageChannel resultsChannel = channelResolver.resolveDestination(integrationResultDestination);
-
-        // complete cloud connector tasks
-        integrationRequestedEvents.stream()
-                                  .map(request -> {
-                                      return new IntegrationResultImpl(new IntegrationRequestImpl(request.getEntity()),
-                                                                       request.getEntity());
-                                  })
-                                  .map(payload -> MessageBuilder.withPayload(payload)
-                                                                .build())
-                                  .forEach(resultsChannel::send);
-
-
-        //then
-        await()
-            .untilAsserted(() -> {
-            List<CloudRuntimeEvent<?, ?>> receivedEvents = streamHandler.getAllReceivedEvents();
-            receivedEvents = receivedEvents.stream()
-                    .filter(event -> startProcessEntity.getBody().getId().equals(event.getProcessInstanceId()))
-                    .collect(Collectors.toList());
-
-            List<CloudIntegrationResultReceivedEvent> receivedIntegrationResultEvents = receivedEvents.stream()
-                    .filter(event -> event.getEventType() == INTEGRATION_RESULT_RECEIVED)
-                    .map(CloudIntegrationResultReceivedEvent.class::cast)
-                    .collect(Collectors.toList());
-
-            assertThat(receivedIntegrationResultEvents)
-                    .extracting(RuntimeEvent::getEventType, event -> ((IntegrationContext) event.getEntity()).getClientId())
-                    .containsExactlyInAnyOrder(
-                            tuple(INTEGRATION_RESULT_RECEIVED, "performBusinessTaskCloudConnector")
                     );
 
             assertThat(receivedEvents).extracting(RuntimeEvent::getEventType,
